pyproject.toml
[project]
name = "django-financial"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.5",
    "django-restframework>=0.0.1",
    "drf-spectacular>=0.28.0",
    "drf-yasg>=1.21.10",
    "gunicorn>=21.2",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "uvicorn>=0.24.0",
    "python-dotenv>=0.21.0",
    "django-filter>=25.1",
    "djangorestframework-simplejwt>=5.5.1",

]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "ruff>=0.12.10",
]
